Rename file and deploy on playground
